+++ packages/apipgen-parser-swagger/src/transform/parameters.ts
@@ -6,14 +6,22 @@ export interface ParameterTransformOptions {
   description: string[]
   responseType: string
   syntax: 'typescript' | 'ecmascript'
-  returnType?: boolean
+  generic?: string
 }
 export function transformParameters(parameters: StatementField[], options: ParameterTransformOptions) {
-  const { configRead, syntax, interfaces, description, responseType, returnType = true } = options
-  const typeImport = configRead.outputs.find(v => v.type === 'typings')?.import
-  const isGenerateType = configRead.outputs.map(v => v.type).includes('typings')
-  const TypeNamespace = syntax === 'ecmascript' ? `import('${typeImport}')` : 'Types'
-  const spaceResponseType = `${TypeNamespace}.Response<${spliceTypeSpace(responseType)}>`
+  const { configRead, syntax, interfaces, description, responseType } = options
+  const importType = configRead.outputs.find(v => v.type === 'typings')?.import
+  const isGenerate = configRead.outputs.map(v => v.type).includes('typings')
+  const namespace = syntax === 'ecmascript' ? `import('${importType}')` : 'Types'
+  const infer = configRead.graphs.response.infer || ''
+  const generic = parseGenericType(configRead.graphs.response.generic || options.generic, syntax)
+  const spaceResponseType = parseResponseType({
+    responseType: spliceTypeSpace(responseType),
+    generic,
+    infer,
+    namespace,
+  })
+
   for (const parameter of parameters || []) {
     if (!parameter.type)
       continue
@@ -21,14 +29,14 @@ export function transformParameters(parameters: StatementField[], options: Param
 
     if (syntax !== 'ecmascript')
       continue
-    if (isGenerateType)
+    if (isGenerate)
       description.push(`@param {${parameter.type}${parameter.required ? '' : '='}} ${parameter.name}`)
     parameter.type = undefined
     parameter.required = true
   }
 
-  if (isGenerateType && syntax === 'ecmascript' && returnType)
-    description.push(`@return {Promise<${spaceResponseType}>}`)
+  if (isGenerate && syntax === 'ecmascript')
+    description.push(`@return {${spaceResponseType}}`)
 
   function splitTypeSpaces(name: string) {
     const _name = name
@@ -45,9 +53,24 @@ export function transformParameters(parameters: StatementField[], options: Param
       return splitTypeSpaces(name)
     if (name.includes('|') && name.includes('[]'))
       return `${splitTypeSpaces(name)}[]`
-    const someType = interfaces.map(v => v.name).includes(name.replace('[]', ''))
-    return (isGenerateType && someType) ? `${TypeNamespace}.${name}` : name
+    const isExists = interfaces.map(v => v.name).includes(name.replace('[]', ''))
+    return (isGenerate && isExists) ? `${namespace}.${name}` : name
   }
 
   return { spaceResponseType }
+}
+
+function parseGenericType(generic = '', syntax: 'typescript' | 'ecmascript') {
+  if (!generic)
+    generic = syntax === 'ecmascript' ? 'Promise<{__type__}>' : '{__type__}'
+  return generic
+}
+
+function parseResponseType({ responseType = '', namespace = '', generic = '', infer = '' }) {
+  if (infer)
+    responseType = `${namespace}.Infer<${responseType}>`
+
+  responseType = generic.replace('{__type__}', responseType)
+
+  return responseType
 }