+++ packages/apipgen-pipeline/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @apipgen/pipeline
 
+## 0.1.7
+
+### Patch Changes
+
+- Merge branch 'main' of https://github.com/TuiMao233/apipgen
+- Updated dependencies
+  - apipgen@0.3.13
+
+## 0.1.6
+
+### Patch Changes
+
+- 🦋 apipgen@0.3.10
+- Updated dependencies
+  - apipgen@0.3.11
+
 ## 0.1.5
 
 ### Patch Changes